+++ setup/config.js
@@ -1,0 +1,13 @@
+export const FILTER_FIELDS = {
+  candidate_gender: {},
+  state: {},
+  amt_raised_diff_pct_2018: {
+    num_buckets: 5,
+  },
+}
+
+export const BREAK_CHAR = ',,,'
+
+export const SCATTERPLOT_X_FIELD = 'amt_raised_diff_2020'
+
+export const SCATTERPLOT_Y_FIELD = 'vote_diff_2018'
