+++ utils/utils.js
@@ -1,0 +1,4 @@
+export const dollarToFloat = (dollar) =>
+  Number.parseFloat(dollar.replace('$', '').replace(',', ''))
+
+export const stringNumToFloat = (str) => Number.parseFloat(str.replace(',', ''))
